+++ src/pages/MyCareer/ProfileEditPage/ProfileEditPage.jsx
@@ -1,21 +1,14 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import UnderlineButton from '../../../components/common/Button/UnderlineButton/UnderlineButton';
 import ProfileSetting from '../../../components/common/ProfileSetting/ProfileSetting';
 import * as S from './styled/styled';
 import AccountPopUp from '../../../components/common/Popups/AccountPopUp/AccountPopUp';
-import { useProfile } from '../../../hooks/useProfile';
-import { useAuthStore } from '../../../store/authStore';
-import { useDeleteProfile, useEditProfile } from '../../../apis/Profile/useProfileApi';
-import { useNavigate } from 'react-router-dom';
+import { useProfileEdit, useDeleteAccount } from './useProfileEdit';
 
 const ProfileEditPage = () => {
     const [isPopUp, setIsPopUp] = useState(false);
-    const [isDeleting, setIsDeleting] = useState(false);
-    const { logout, user } = useAuthStore();
-    const { canSave, emailError, profile, handleProfileFieldChange } = useProfile();
-    const { mutate: deleteProfile } = useDeleteProfile();
-    const { mutate: modifyProfile } = useEditProfile(profile);
-    const navigate = useNavigate();
+    const { profile, handleProfileFieldChange, handleSave } = useProfileEdit();
+    const { isDeleting, handleDeleteUser } = useDeleteAccount();
 
     const handlePopUpOpen = () => {
         setIsPopUp(true);
@@ -23,28 +16,6 @@ const ProfileEditPage = () => {
 
     const handlePopUpClose = () => {
         setIsPopUp(false);
-    };
-
-    const handleSave = () => {
-        if (!canSave) {
-            alert('항목을 모두 입력해주세요!');
-            return;
-        }
-
-        if (emailError) {
-            alert('유효한 이메일 주소를 입력하세요.');
-            return;
-        }
-        modifyProfile(profile);
-    };
-
-    const handleDeleteUser = () => {
-        setIsDeleting(true);
-        //deleteProfile();
-        setTimeout(() => {
-            logout();
-            navigate('/');
-        }, 200);
     };
 
     return (
